+++ .github/labeler.yml
@@ -1,3 +1,3 @@
 themes: themes/index.js
 doc-translation: docs/*
-card-i18n: src/translations.js
+card-i18n: src/assets/translations.js
