+++ api/top-langs.js
@@ -9,7 +9,7 @@ import {
   renderError,
 } from "../src/common/utils.js";
 import { fetchTopLanguages } from "../src/fetchers/top-languages-fetcher.js";
-import { isLocaleAvailable } from "../src/translations.js";
+import { isLocaleAvailable } from "../src/assets/translations.js";
 
 dotenv.config();
