+++ api/wakatime.js
@@ -8,7 +8,7 @@ import {
   renderError,
 } from "../src/common/utils.js";
 import { fetchWakatimeStats } from "../src/fetchers/wakatime-fetcher.js";
-import { isLocaleAvailable } from "../src/translations.js";
+import { isLocaleAvailable } from "../src/assets/translations.js";
 
 dotenv.config();
