+++ src/assets/translations.js
@@ -1,4 +1,4 @@
-import { encodeHTML } from "./common/utils.js";
+import { encodeHTML } from "../common/utils.js";
 
 /**
  * Retrieves stat card labels in the available locales.
