+++ src/cards/top-languages-card.js
@@ -10,7 +10,7 @@ import {
   lowercaseTrim,
   measureText,
 } from "../common/utils.js";
-import { langCardLocales } from "../translations.js";
+import { langCardLocales } from "../assets/translations.js";
 
 const DEFAULT_CARD_WIDTH = 300;
 const MIN_CARD_WIDTH = 230;
