+++ src/cards/wakatime-card.js
@@ -8,8 +8,8 @@ import {
   getCardColors,
   lowercaseTrim,
 } from "../common/utils.js";
-import { getStyles } from "../getStyles.js";
-import { wakatimeCardLocales } from "../translations.js";
+import { getStyles } from "../common/getStyles.js";
+import { wakatimeCardLocales } from "../assets/translations.js";
 
 /** Import language colors.
  *
